move directory and change implementation
