+++ compiler/luci/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_subdirectory(env)
 add_subdirectory(log)
 add_subdirectory(lang)
+add_subdirectory(profile)
 add_subdirectory(service)
 add_subdirectory(pass)
 add_subdirectory(logex)
