+++ compiler/luci/profile/README.md
@@ -1,0 +1,3 @@
+# luci-profile
+
+`luci-profile` provides profiling related items.
